refactor(exchange): 교환 패키지 구조 변경
