+++ src/main/java/com/musinsa/orders/domain/exchange/ExchangePolicy.java
@@ -1,5 +1,8 @@
-package com.musinsa.orders.domain;
+package com.musinsa.orders.domain.exchange;
 
+import com.musinsa.orders.domain.Money;
+import com.musinsa.orders.domain.Order;
+import com.musinsa.orders.domain.OrderLineItem;
 import java.util.List;
 
 public class ExchangePolicy {
